UPDATE extract plugin into own file UPDATE log
